Drop JS-alias emulation (plan §6)
- barkify: xof/yof offsets + write_back flag replace the pre-shifted
  brCopy copies in tree04; the jittered-outline leak now applies only
  to the main trunk
- flatMount: closing point synced via front() = back() with intent
- mountplanner: planmtx is map<int,int>; a materialize-the-slab pass
  plus find-guarded span marking express the old NaN semantics as
  'absent cell = unoccupied'
- Man cloth reverse documented as deliberate outline orientation;
  'JS: float' bound comments rewritten as intent

Output verified byte-identical to the previous build on host and wasm.

+++ src/scene/scene.h
@@ -20,7 +20,7 @@ struct Scene {
 	double cursx = 0;
 	double lasttick = 0;
 	double windx = 3000, windy = 800;
-	std::map<int, double> planmtx; // JS: sparse array with NaN semantics
+	std::map<int, int> planmtx; // mount planner occupancy grid
 
 	void seed(const std::string& s);
 	void reset();
